backend/pyproject.toml
[project]
name = "opc-backend"
version = "0.1.0"
description = "Open Property Core — Django backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0,<6",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "drf-spectacular>=0.27",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    "gunicorn>=22.0",
    "celery>=5.3",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "mypy>=1.8",
    "flake8>=7.0",
    "isort>=5.13",
    "django-stubs>=5.0",
    "pytest>=8.0",
    "pytest-django>=4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '\.venv|\.git|__pycache__|migrations'

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["opc"]
skip = [".venv", "migrations"]
src_paths = ["."]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "opc.settings"

[tool.pytest.ini_options]
python_files = ["test_*.py"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = "-v"
